# cmd/stationd/docs/swagger.yaml
basePath: /feeStation/api
definitions:
  station_handlers.PoolInfo:
    properties:
      decimals:
        type: integer
      poolAddress:
        description: base58,bech32 or hex
        type: string
      swapRate:
        description: decimals 6
        type: string
      symbol:
        type: string
    type: object
  station_handlers.ReqPostSwapInfo:
    properties:
      inAmount:
        description: decimal
        type: string
      minOutAmount:
        description: decimal
        type: string
      poolAddress:
        type: string
      stafihubAddress:
        description: hex
        type: string
      symbol:
        type: string
    type: object
  station_handlers.RspGetSwapInfo:
    properties:
      swapStatus:
        type: integer
    type: object
  station_handlers.RspPoolInfo:
    properties:
      poolInfoList:
        items:
          $ref: '#/definitions/station_handlers.PoolInfo'
        type: array
      swapMaxLimit:
        description: decimals 6
        type: string
      swapMinLimit:
        description: decimals 6
        type: string
    type: object
  station_handlers.RspPostSwapInfo:
    properties:
      uuid:
        type: string
    type: object
  utils.Rsp:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8083
info:
  contact:
    email: tpkeeper@qq.com
    name: tk
  description: feeStation api document.
  title: feeStation API
  version: "1.0"
paths:
  /v1/station/poolInfo:
    get:
      description: get pool info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Rsp'
            - properties:
                data:
                  $ref: '#/definitions/station_handlers.RspPoolInfo'
              type: object
      summary: get pool info
      tags:
      - v1
  /v1/station/swapInfo:
    get:
      description: get swap info
      parameters:
      - description: uuid hex string
        in: query
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Rsp'
            - properties:
                data:
                  $ref: '#/definitions/station_handlers.RspGetSwapInfo'
              type: object
      summary: get swap info
      tags:
      - v1
    post:
      consumes:
      - application/json
      description: post swap info
      parameters:
      - description: user swap info
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/station_handlers.ReqPostSwapInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Rsp'
            - properties:
                data:
                  $ref: '#/definitions/station_handlers.RspPostSwapInfo'
              type: object
      summary: post swap info
      tags:
      - v1
swagger: "2.0"